StudentSage/demoquiz-main/requirements.txt
Django==5.0.6
faiss-cpu==1.8.0
google-generativeai==0.5.4
langchain==0.2.5
langchain-community==0.2.5
pydantic==2.7.4
pydantic_core==2.18.4
pypdf==4.2.0
PyPDF2==3.0.1
requests==2.32.3